Force int

+++ src/pygor/classes/strf_data.py
@@ -700,7 +700,7 @@ class STRF(Core):
             if isinstance(roi, int):
                 iterate_through = [roi]
             elif isinstance(roi, np.int_):
-                iterate_through = [roi]
+                iterate_through = [roi.astype(int)]
             elif isinstance(roi, Iterable):
                 iterate_through = roi
             else:
